+++ tests/functional_tests/team_ctx_primitive_tester.cpp
@@ -120,9 +120,9 @@ void TeamCtxPrimitiveTester::verifyResults(uint64_t size) {
       (_type == TeamCtxGetTestType || _type == TeamCtxGetNBITestType) ? 0 : 1;
 
   if (args.myid == check_id) {
-    for (int i = 0; i < size; i++) {
+    for (uint64_t i = 0; i < size; i++) {
       if (r_buf[i] != '0') {
-        fprintf(stderr, "Data validation error at idx %d\n", i);
+        fprintf(stderr, "Data validation error at idx %lu\n", i);
         fprintf(stderr, "Got %c, Expected %c\n", r_buf[i], '0');
         exit(-1);
       }
